+++ app/Iconsets/Iconset_Base.php
@@ -227,7 +227,7 @@ class Iconset_Base {
 	 * Get style for given file-type.
 	 *
 	 * @param int    $post_id ID of the icon-post.
-	 * @param string $term_slug ID of the iconset-term.
+	 * @param string $term_slug Slug of the iconset-term.
 	 * @param string $filetype Name for the filetype to add.
 	 * @return string
 	 */